pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "trading-research-lab"
version = "0.1.0"
description = "Laboratorio de investigación y backtesting de estrategias de trading para fondeo tipo FTMO"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "ta-lib",
    "pandas-ta>=0.3",
    "yfinance>=0.2",
    "requests>=2.31",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "PyYAML>=6.0",
    "pyarrow>=12.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "plotly>=5.15",
    "rich>=13.0",
    "loguru>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-mock>=3.11",
    "ruff>=0.1",
    "mypy>=1.5",
    "ipykernel>=6.25",
    "jupyterlab>=4.0",
]
ml = [
    "lightgbm>=4.0",
    "xgboost>=1.7",
    "optuna>=3.3",
    "shap>=0.42",
]
lean = [
    "lean>=1.0",
]
freqtrade = [
    "freqtrade>=2024.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"
